Task2: Customer id changed to Integer

+++ src/main/java/com/cinema/cinemaapp/repository/CustomerRepository.java
@@ -2,8 +2,8 @@ package com.cinema.cinemaapp.repository;
 
 
 
-public interface CustomerRepository extends CrudRepository<Long, Customer> {
-    Customer deleteCustomerById(Long id);
+public interface CustomerRepository extends CrudRepository<Integer, Customer> {
+    Customer deleteCustomerById(Integer id);
     Customer findCustomerById(Integer id);
     Customer findCustomerByEmail(String email);
     List<Customer> findAll();
